+++ core/core/src/main/kotlin/edu/wpi/axon/core/dsl/variable/InferenceSession.kt
@@ -1,7 +1,7 @@
 package edu.wpi.axon.core.dsl.variable
 
 import edu.wpi.axon.core.dsl.Import
-import edu.wpi.axon.core.dsl.PathValidator
+import edu.wpi.axon.core.dsl.validator.path.PathValidator
 import org.koin.core.KoinComponent
 import org.koin.core.inject
